Renamed k8sDeployment to k8sObject
